# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.APIResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  models.CreateUserByAdminRequest:
    properties:
      commission_percentage:
        maximum: 100
        minimum: 1
        type: integer
      email:
        type: string
      first_name:
        type: string
      institution_id:
        type: string
      last_name:
        type: string
      national_id:
        type: string
      passport_number:
        type: string
      password:
        maxLength: 16
        minLength: 5
        type: string
      phone:
        type: string
      profile_picture:
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.UserRole'
        enum:
        - member
        - admin
        - trainer
        - instructor
        - frontdesk
        - finance
        - seler
      seler_type:
        allOf:
        - $ref: '#/definitions/models.SelerType'
        enum:
        - seler
        - promoter
      specialization:
        type: string
      username:
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - role
    type: object
  models.CreateUserRequest:
    properties:
      client_id:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      national_id:
        type: string
      passport_number:
        type: string
      password:
        maxLength: 16
        minLength: 5
        type: string
      phone:
        type: string
      role:
        $ref: '#/definitions/models.UserRole'
      status:
        $ref: '#/definitions/models.UserStatus'
      username:
        type: string
    required:
    - client_id
    - email
    - first_name
    - last_name
    - password
    - phone
    type: object
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.PaginatedResponse:
    properties:
      current_page:
        type: integer
      last_page:
        type: integer
      list: {}
      next_page:
        type: integer
      previous_page:
        type: integer
      status:
        type: string
      total:
        type: integer
    type: object
...
  models.PasswordUpdateRequest:
    properties:
      new_password:
        minLength: 6
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
  models.SelerType:
    enum:
    - seler
    - promoter
    type: string
    x-enum-varnames:
    - SelerTypeSeler
    - SelerTypePromoter
  models.SubscriptionStatus:
    enum:
    - active
    - inactive
    - expired
    - onhold
    - paused
    - canceled
    type: string
    x-enum-varnames:
    - SubscriptionActive
    - SubscriptionInactive
    - SubscriptionExpired
    - SubscriptionOnHold
    - SubscriptionPaused
    - SubscriptionCanceled
  models.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  models.UpdateUserRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      national_id:
        type: string
      notes:
        type: string
      passport_number:
        type: string
      password:
        maxLength: 16
        minLength: 5
        type: string
      phone:
        type: string
      profile_picture:
        type: string
      role:
        $ref: '#/definitions/models.UserRole'
      specialization:
        type: string
      status:
        $ref: '#/definitions/models.UserStatus'
      username:
        type: string
    type: object
  models.UserResponse:
    properties:
      client_id:
        type: string
      commission_percentage:
        type: integer
      created_at:
        type: string
      created_by:
        type: string
      email:
        type: string
      first_name:
        type: string
      has_active_subscription:
        type: boolean
      id:
        type: string
      institution_id:
        type: string
      is_active:
        type: boolean
      last_name:
        type: string
      national_id:
        type: string
      notes:
        type: string
      otp_required:
        type: boolean
      passport_number:
        type: string
      phone:
        type: string
      profile_picture:
        type: string
      referral_code:
        type: string
      role:
        $ref: '#/definitions/models.UserRole'
      seler_type:
        $ref: '#/definitions/models.SelerType'
      slug:
        type: string
      specialization:
        type: string
      status:
        $ref: '#/definitions/models.UserStatus'
      subscription_status:
        $ref: '#/definitions/models.SubscriptionStatus'
      updated_at:
        type: string
      username:
        type: string
    type: object
  models.UserRole:
    enum:
    - user
    - admin
    - super_admin
    - trainer
    - instructor
    - frontdesk
    - finance
    - seler
    - member
    type: string
    x-enum-varnames:
    - RoleUser
    - RoleAdmin
    - RoleSuperAdmin
    - RoleTrainer
    - RoleInstructor
    - RoleFrontdesk
    - RoleFinance
    - RoleSeler
    - RoleMember
  models.UserStatus:
    enum:
    - active
    - inactive
    - suspended
    - deleted
    type: string
    x-enum-varnames:
    - StatusActive
    - StatusInactive
    - StatusSuspended
    - StatusDeleted
host: localhost:8082
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive user management system with authentication and authorization
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: User Management API
  version: "1.0"
paths:
...
  /users:
    get:
      description: Get paginated list of users with optional filtering
      parameters:
      - default: 1
        description: Page number
        in: query
        name: pageNumber
        type: integer
      - default: 10
        description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Filter by role
        in: query
        name: role
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Search in email, username, first_name, last_name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get All Users
      tags:
      - Users
  /users/{id}:
    delete:
      description: Soft delete a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete User
      tags:
      - Users
    get:
      description: Get a single user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get User by ID
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Update user information
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User update data
        in: body
        name: updateUserRequest
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update User
      tags:
      - Users
...
  /users/register:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: User registration data
        in: body
        name: createUserRequest
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Register User
      tags:
      - Users
  /users/registerusersbyadmin:
    post:
      consumes:
      - application/json
      description: Create a new user by admin with extended fields
      parameters:
      - description: Admin user creation data
        in: body
        name: createUserByAdminRequest
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserByAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create User by Admin
      tags:
      - Users
  /users/slug/{slug}:
    get:
      description: Get a single user by their slug
      parameters:
      - description: User Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get User by Slug
      tags:
      - Users
  /users/update-password:
    post:
      consumes:
      - application/json
      description: Update user's own password
      parameters:
      - description: Password update data
        in: body
        name: passwordUpdateRequest
        required: true
        schema:
          $ref: '#/definitions/models.PasswordUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Password
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"